+++ sillo/__init__.py
@@ -110,7 +110,7 @@ Common Patterns:
 
 from sillo.core.routing import Group, Route, Router, WebsocketRoute
 
-__version__: str = "0.3.0"
+__version__: str = "0.3.1"
 
 from sillo.core.dependencies import Depend